Commit 6.0: Updated all my layers of concern. Changed my NPC entity's field objects to Integer instead of Long, and all corresponding coding, which fixed my null values for my Npc table FKs

+++ src/main/java/npc/dao/DefaultBackgroundDao.java
@@ -29,7 +29,7 @@ public class DefaultBackgroundDao implements BackgroundDao {
         
         // @formatter:off
         return Background.builder()
-            .backgroundPk(rs.getLong("background_pk"))
+            .backgroundPk(rs.getInt("background_pk"))
             .backgroundId(BackgroundType.valueOf(rs.getString("background_id")))
             .build();
       }
